adLastResources

+++ src/main/java/com/wheelcredit/Backend/model/SmartPayment.java
@@ -19,22 +19,22 @@ public class SmartPayment {
     private Long id;
 
     @Column(name="selling_price_ asset",nullable=false)
-    private float sellingPriceAsset; //precio de venta activo
+    private double sellingPriceAsset; //precio de venta activo
 
     @Column(name="payment_plan_type",nullable=false)
     private int paymentPlanType;//tipo de plan de pago
 
     @Column(name="initial_installment",nullable=false)
-    private float initialInstallment;// cuota inicial
+    private double initialInstallment;// cuota inicial
 
     @Column(name = "final_fee", nullable = false)//
-    private float finalFee;// cuota final
+    private double finalFee;// cuota final
 
     @Column(name = "number_years", nullable = false)//
     private int numberOfYears;// numero de años
 
     @Column(name= "interest_rate",nullable = false)
-    private float interestRate; //tasa de interes
+    private double interestRate; //tasa de interes
 
     @Column(name= "interest_type",nullable = false)
     private String interestType; //tipo de tasa
@@ -46,37 +46,37 @@ public class SmartPayment {
     private int paymentFrequency;  // frecuencia de pago
 
     @Column(name= "notary_costs",nullable = false)
-    private float notaryCosts; //costos notariales
+    private double notaryCosts; //costos notariales
 
     @Column(name= "registration_costs",nullable = false)
-    private float registrationCosts; // costos de registro
+    private double registrationCosts; // costos de registro
 
     @Column(name= "appraisal",nullable = false)
     private float appraisal; // tasacion
 
     @Column(name= "study_commission",nullable = false)
-    private float studyCommission; //comision de estudio
+    private double studyCommission; //comision de estudio
 
     @Column(name= "activation_commission",nullable = false)
-    private float activationCommission; // comision de activacion
+    private double activationCommission; // comision de activacion
 
     @Column(name= "gps",nullable = false)
-    private float gps;
+    private double gps;
 
     @Column(name= "shipping_costs",nullable = false)
-    private float shippingCosts; //portes
+    private double shippingCosts; //portes
 
     @Column(name= "administrative_expenses",nullable = false)
-    private float administrativeExpenses; //gastos administrativos
+    private double administrativeExpenses; //gastos administrativos
 
     @Column(name= "life_insurance",nullable = false)
-    private float lifeInsurance; //seguro desagraven
+    private double lifeInsurance; //seguro desagraven
 
     @Column(name= "risk_insurance",nullable = false)
-    private float riskInsurance; // seguro de riesgo
+    private double riskInsurance; // seguro de riesgo
 
     @Column(name= "discount_rate",nullable = false)
-    private float discountRate; //tasade descuento
+    private double discountRate; //tasade descuento
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="client_id", referencedColumnName = "id")
